# docs/openapi.yaml
openapi: 3.0.3
info:
  title: BitByBit Habits API
  description: |
    API REST de BitByBit, un habit tracker que recompensa completar tareas con sats reales via Lightning Network.

    **Caso de uso MVP**: "Mesada 2.0" donde sponsors (padres/tutores) crean habitos con recompensa en sats
    y kids los completan para ganar.

    ## Autenticacion

    La API usa autenticacion basada en cookies de sesion (`session`). Para autenticarse:

    1. Registrar un usuario con `POST /api/auth/register`
    2. Iniciar sesion con `POST /api/auth/login`
    3. La cookie `session` se setea automaticamente en la respuesta

    Todos los endpoints (excepto auth) requieren una sesion activa.

    ## Formato de respuesta

    Todas las respuestas siguen el formato `ApiResponse`:
    ```json
    {
      "success": true,
      "data": { ... }
    }
    ```
    En caso de error:
    ```json
    {
      "success": false,
      "error": "Descripcion del error"
    }
    ```
  version: 1.0.0
  contact:
    name: BitByBit Team
servers:
  - url: /api
    description: Servidor de la API
...
paths:
  /auth/register:
    post:
      tags:
        - Auth
      summary: Registrar un nuevo usuario
      description: |
        Crea un nuevo usuario en el sistema. El email y username deben ser unicos.
        El rol puede ser `sponsor` (padre/tutor) o `kid` (hijo/a).
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
            example:
              email: "padre@ejemplo.com"
              username: "papajuan"
              password: "mipassword123"
              display_name: "Juan"
              role: "sponsor"
              locale: "es"
      responses:
        "200":
          description: Usuario registrado exitosamente
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/UserPublic"
              example:
                success: true
                data:
                  id: "550e8400-e29b-41d4-a716-446655440000"
                  email: "padre@ejemplo.com"
                  username: "papajuan"
                  display_name: "Juan"
                  role: "sponsor"
                  locale: "es"
                  created_at: "2026-03-07T10:00:00.000Z"
        "400":
          description: Faltan campos requeridos o rol invalido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"
              examples:
                camposFaltantes:
                  summary: Campos requeridos faltantes
                  value:
                    success: false
                    error: "Faltan campos requeridos"
                rolInvalido:
                  summary: Rol no valido
                  value:
                    success: false
                    error: "Rol invalido"
        "409":
          description: El email o nombre de usuario ya existe
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"
              example:
                success: false
                error: "El email o nombre de usuario ya existe"
        "500":
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"
  /auth/login:
    post:
      tags:
        - Auth
      summary: Iniciar sesion
      description: |
        Autentica un usuario con email o username y password.
        Setea una cookie `session` httpOnly con duracion de 7 dias.
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
            example:
              login: "papajuan"
              password: "mipassword123"
      responses:
        "200":
          description: Sesion iniciada exitosamente. Se setea cookie `session`.
          headers:
            Set-Cookie:
              description: Cookie de sesion httpOnly
              schema:
                type: string
                example: "session=eyJhbGciOi...; HttpOnly; Path=/; SameSite=Lax; Max-Age=604800"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/AuthSession"
              example:
                success: true
                data:
                  user_id: "550e8400-e29b-41d4-a716-446655440000"
                  email: "padre@ejemplo.com"
                  username: "papajuan"
                  display_name: "Juan"
                  role: "sponsor"
                  locale: "es"
        "400":
          description: Faltan campos requeridos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"
              example:
                success: false
                error: "Faltan campos requeridos"
        "401":
          description: Credenciales invalidas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"
              example:
                success: false
                error: "Credenciales invalidas"
        "500":
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"
...
  /auth/session:
    get:
      tags:
        - Auth
      summary: Obtener sesion actual
      description: Devuelve los datos de la sesion del usuario autenticado.
      operationId: getSession
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Sesion activa encontrada
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/AuthSession"
              example:
                success: true
                data:
                  user_id: "550e8400-e29b-41d4-a716-446655440000"
                  email: "padre@ejemplo.com"
                  username: "papajuan"
                  display_name: "Juan"
                  role: "sponsor"
                  locale: "es"
        "401":
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"
              example:
                success: false
                error: "No autenticado"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
      description: Cookie de sesion httpOnly seteada por el endpoint de login
  schemas:
...
    RegisterRequest:
      type: object
      required:
        - email
        - username
        - password
        - display_name
        - role
      properties:
        email:
          type: string
          format: email
          description: Email del usuario (se convierte a minusculas)
        username:
          type: string
          description: Nombre de usuario unico (se convierte a minusculas)
        password:
          type: string
          format: password
          description: Contrasena del usuario
        display_name:
          type: string
          description: Nombre para mostrar
        role:
          type: string
          enum:
            - sponsor
            - kid
          description: Rol del usuario (sponsor = padre/tutor, kid = hijo/a)
        locale:
          type: string
          enum:
            - es
            - en
          default: es
          description: Idioma preferido del usuario
    LoginRequest:
      type: object
      required:
        - login
        - password
      properties:
        login:
          type: string
          description: Email o nombre de usuario (case-insensitive)
        password:
          type: string
          format: password
          description: Contrasena del usuario
...
    ApiResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          description: Indica si la operacion fue exitosa
        data:
          description: Datos de la respuesta (presente cuando success es true)
        error:
          type: string
          description: Mensaje de error (presente cuando success es false)
    ApiResponseError:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          enum:
            - false
        error:
          type: string
          description: Mensaje de error descriptivo
...
    UserPublic:
      type: object
      description: Datos publicos del usuario (sin password_hash)
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        username:
          type: string
        display_name:
          type: string
        role:
          type: string
          enum:
            - sponsor
            - kid
        locale:
          type: string
          enum:
            - es
            - en
        created_at:
          type: string
          format: date-time
    AuthSession:
      type: object
      description: Datos de la sesion del usuario autenticado
      properties:
        user_id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        username:
          type: string
        display_name:
          type: string
        role:
          type: string
          enum:
            - sponsor
            - kid
        locale:
          type: string
          enum:
            - es
            - en
...